+++ Source/Modules/ISHServer/Get-ISHCD.ps1
@@ -42,8 +42,26 @@ function Get-ISHCD
         [string]$SecretKey,
         [Parameter(Mandatory=$false,ParameterSetName="From AWS S3")]
         [string]$SessionToken,
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure FileStorage")]
+        [string]$ShareName,
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure FileStorage")]
+        [ValidatePattern(".+\.[0-9]+\.0\.[0-9]+\.[0-9]+.*\.exe")]
+        [string]$Path,
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure BlobStorage")]
+        [string]$ContainerName,
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure BlobStorage")]
+        [ValidatePattern(".+\.[0-9]+\.0\.[0-9]+\.[0-9]+.*\.exe")]
+        [string]$BlobName,
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure FileStorage")]
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure BlobStorage")]
+        [string]$StorageAccountName,
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure FileStorage")]
+        [Parameter(Mandatory=$true,ParameterSetName="From Azure BlobStorage")]
+        [string]$StorageAccountKey,
         [Parameter(Mandatory=$false,ParameterSetName="From FTP")]
         [Parameter(Mandatory=$false,ParameterSetName="From AWS S3")]
+        [Parameter(Mandatory=$false,ParameterSetName="From Azure FileStorage")]
+        [Parameter(Mandatory=$false,ParameterSetName="From Azure BlobStorage")]
         [switch]$Expand=$false,
         [Parameter(Mandatory=$true,ParameterSetName="List")]
         [switch]$ListAvailable
@@ -90,6 +108,40 @@ function Get-ISHCD
                 }
 
                 $newItem=Get-ISHS3Object -Key $Key @hash
+                if($Expand)
+                {
+                    . $PSScriptRoot\Expand-ISHCD.ps1
+                    Expand-ISHCD -FileName $newItem.Name
+                }
+                break        
+            }
+            'From Azure FileStorage' {
+                . $PSScriptRoot\Private\Get-ISHAzureFileObject.ps1
+                $hash=@{
+                    ShareName=$ShareName
+                    LocalFolder=$localPath
+                    StorageAccountName=$StorageAccountName
+                    StorageAccountKey=$StorageAccountKey
+                }
+
+                $newItem=Get-ISHAzureFileObject -Path $Path @hash
+                if($Expand)
+                {
+                    . $PSScriptRoot\Expand-ISHCD.ps1
+                    Expand-ISHCD -FileName $newItem.Name
+                }
+                break
+            }
+            'From Azure BlobStorage' {
+                . $PSScriptRoot\Private\Get-ISHAzureBlobObject.ps1
+                $hash=@{
+                    ContainerName=$ContainerName
+                    LocalFolder=$localPath
+                    StorageAccountName=$StorageAccountName
+                    StorageAccountKey=$StorageAccountKey
+                }
+
+                $newItem=Get-ISHAzureBlobObject -BlobName $BlobName @hash
                 if($Expand)
                 {
                     . $PSScriptRoot\Expand-ISHCD.ps1