+++ Source/Modules/ISHServer/metadata.ps1
@@ -15,5 +15,5 @@
 #>
 
 <#PSScriptInfo
-.VERSION 1.1
+.VERSION 1.2
 #>